Fix key of HashMap Memory and IODevice banks Fix set all half-rows of keyboard to non-pressed Fix opCodes table for missing PUSH_rp2_p entries Add test for missing PUSH_rp2_p: Pass Fix initial value for SP (TODO general method) TODO: problems with BitSet when all bits are 0: toByteArray()[0] throws java.lang.ArrayIndexOutOfBoundsException

+++ src/main/java/com/github/jagarsoft/GUI/ZXSpectrumEmulator.java
@@ -100,7 +100,7 @@ public class ZXSpectrumEmulator {
         spectrum.addMemory(0x4000, new VRAM(screen));
         spectrum.addMemory(0x8000, new RAMMemory(16 * 1024));
         /*spectrum.addMemory(0xC000, new ROMMemory(16 * 1024))*/
-        spectrum.addIODevice(0xFE, keyboard); // new ZXSpectrumIO(new ZXSpectrumKeyboard(), new ZXSpectrumBeeperTapeAndBorder())
+        spectrum.addIODevice((short)0xFE, keyboard); // new ZXSpectrumIO(new ZXSpectrumKeyboard(), new ZXSpectrumBeeperTapeAndBorder())
 
         spectrum.reset();
